fix(cmd): small sets of fixes in cmd package (#170)
* fix(typo): malformated struct tag gnokeys.go

* fix(type_conversion): remove redundant type conversion

* fix(composite): remove redundant type

* fix(type_omitted): remove unnecessary type declaration

+++ cmd/gnotxport/export.go
@@ -47,7 +47,7 @@ func txExportApp(cmd *command.Command, args []string, iopts interface{}) error {
 		return err
 	}
 
-	for height := int64(opts.StartHeight); height <= last; height++ {
+	for height := opts.StartHeight; height <= last; height++ {
 		block, err := c.Block(&height)
 		if err != nil {
 			panic(err)
